Added arguments, random dungeon population and more
Added arguments for tests, to change some important data from the game, and change styles. Also created a constructor that populates the dungeon with monsters, which will later be passed to a "build level" constructor for easy maintenability. Added behavior for monsters not be drawn if they are in tiles that arent visible. Changed the way _id is created.