Item drop added and targeting problem fixed
now items can be droped by the player on the ground. Also a targeting error was being caused because it was targeting dead monsters and ending in a KeyError when trying to make it "take damage". It was fixed by changing the tags in the dead corpse to ["corpse"], making it not targeted anymore by the target_tag